feat(product): 实现商品BOM管理功能并优化表单数据处理
- 新增商品BOM及BOM明细的删除接口- 新增根据商品ID查询BOM信息接口
- 优化商品创建/更新接口，支持BOM数据同步处理
-修复商品列表排序字段错误问题- 调整商品表单组件数据绑定方式，提升响应性
- 统一数值类型为BigDecimal，提高精度处理
- 完善BOM相关业务逻辑异常处理机制

+++ my-project-backend/src/main/java/com/example/entity/cangku/req/ProductCreateReq.java
@@ -3,6 +3,8 @@ package com.example.entity.cangku.req;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @Author YangJian
  * @Description
@@ -27,9 +29,9 @@ public class ProductCreateReq {
     //出货单位（比如按箱子出货）
     private String outUnitCode;
     //出货单位数量，比如出货时一箱子多少货
-    private Long outUnitPerNum;
+    private BigDecimal outUnitPerNum;
     //单品重量(可选)
-    private Double weightPerUnit;
+    private BigDecimal weightPerUnit;
     //颜色
     private String color;
     //最低库存
