+++ my-project-backend/src/main/java/com/example/entity/cangku/resp/ProductPageListResp.java
@@ -3,6 +3,7 @@ package com.example.entity.cangku.resp;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -32,9 +33,9 @@ public class ProductPageListResp {
     private String outUnitCode;
     private String outUnitName;
     //出货单位数量，比如出货时一箱子多少货
-    private Long outUnitPerNum;
+    private BigDecimal outUnitPerNum;
     //单品重量(可选)
-    private Double weightPerUnit;
+    private BigDecimal weightPerUnit;
     //颜色
     private String color;
     //最低库存
